+++ index.js
@@ -100,6 +100,10 @@ const config = {
     'plugin:import/recommended'
   ],
   rules: {
+    // This rule has too many false positives:
+    // https://github.com/eslint/eslint/issues/11899
+    'require-atomic-updates': 'off',
+
     'require-jsdoc': [
       'error',
       {