+++ changelog.md
@@ -1,5 +1,11 @@
 # eslint-config-env changelog
 
+## Next
+
+### Patch
+
+- Updated dependencies.
+
 ## 8.0.0
 
 ### Major